pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codebase-rag"
version = "0.1.0"
description = "A Retrieval-Augmented Generation system for codebase search and developer assistance"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "developer@example.com"},
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.25.0",
    "chromadb>=0.4.0",
    "faiss-cpu>=1.7.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "tiktoken>=0.4.0",
    "tree-sitter>=0.20.0",
    "pygments>=2.15.0",
    "python-magic>=0.4.27",
    "gitpython>=3.1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"] 